fix: lint errors

+++ contracts/Pool/IOddzLiquidityPool.sol
@@ -9,10 +9,14 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
  * @notice Accumulates liquidity in USD from LPs
  */
 interface IOddzLiquidityPool {
-    struct LockedLiquidity { uint amount; uint premium; bool locked; }
+    struct LockedLiquidity {
+        uint256 amount;
+        uint256 premium;
+        bool locked;
+    }
 
-    event Profit(uint indexed _id, uint _amount);
-    event Loss(uint indexed _id, uint _amount);
+    event Profit(uint256 indexed _id, uint256 _amount);
+    event Loss(uint256 indexed _id, uint256 _amount);
     event Provide(address indexed _account, uint256 _amount, uint256 _writeAmount);
     event Withdraw(address indexed _account, uint256 _amount, uint256 _writeAmount);
 
@@ -48,7 +52,11 @@ interface IOddzLiquidityPool {
      * @param _account Provider account address
      * @param _amount Funds that should be sent
      */
-    function send(uint256 _id, address payable _account, uint256 _amount) external;
+    function send(
+        uint256 _id,
+        address payable _account,
+        uint256 _amount
+    ) external;
 
     /**
      * @notice Returns the amount of USD available for withdrawals
